refactor: convert block-hub into widget
