+++ packages/blocks/src/page-block/edgeless/components/toolbar/note/note-menu-config.ts
@@ -1,10 +1,10 @@
 import type { TemplateResult } from 'lit';
 
+import type { NoteChildrenFlavour } from '../../../../../_common/utils/index.js';
 import {
   BLOCKHUB_LIST_ITEMS,
   BLOCKHUB_TEXT_ITEMS,
-} from '../../../../../_common/components/block-hub.js';
-import type { NoteChildrenFlavour } from '../../../../../_common/utils/index.js';
+} from '../../../../../_common/widgets/block-hub/config.js';
 
 export const BUTTON_GROUP_LENGTH = 10;
