+++ packages/blocks/src/page-block/types.ts
@@ -1,3 +1,4 @@
+import type { AFFINE_BLOCK_HUB_WIDGET } from '../_common/widgets/block-hub/index.js';
 import type { AFFINE_DOC_DRAGGING_AREA_WIDGET } from '../_common/widgets/doc-dragging-area/index.js';
 import type { AFFINE_DOC_REMOTE_SELECTION_WIDGET } from '../_common/widgets/doc-remote-selection/doc-remote-selection.js';
 import type { AFFINE_DRAG_HANDLE_WIDGET } from '../_common/widgets/drag-handle/index.js';
@@ -10,6 +11,7 @@ import type { DocPageBlockComponent } from './doc/doc-page-block.js';
 import type { EdgelessPageBlockComponent } from './edgeless/edgeless-page-block.js';
 
 export type DocPageBlockWidgetName =
+  | typeof AFFINE_BLOCK_HUB_WIDGET
   | typeof AFFINE_MODAL_WIDGET
   | typeof AFFINE_SLASH_MENU_WIDGET
   | typeof AFFINE_LINKED_PAGE_WIDGET
@@ -18,6 +20,7 @@ export type DocPageBlockWidgetName =
   | typeof AFFINE_FORMAT_BAR_WIDGET
   | typeof AFFINE_DOC_REMOTE_SELECTION_WIDGET;
 export type EdgelessPageBlockWidgetName =
+  | typeof AFFINE_BLOCK_HUB_WIDGET
   | typeof AFFINE_MODAL_WIDGET
   | typeof AFFINE_SLASH_MENU_WIDGET
   | typeof AFFINE_LINKED_PAGE_WIDGET
